Feature 20251211 hangyemoban

See merge request pd_str/marketing!1910

+++ marketing-cs/src/main/java/com/br/marketing/mapper/MarketingDataCleanGeneralConfigMapperBase.java
@@ -2,11 +2,12 @@ package com.br.marketing.mapper;
 
 import com.br.marketing.entity.MarketingDataCleanGeneralConfig;
 import com.br.marketing.entity.MarketingDataCleanGeneralConfigExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
+import java.util.List;
+
 public interface MarketingDataCleanGeneralConfigMapperBase {
-    long countByExample(MarketingDataCleanGeneralConfigExample example);
+    int countByExample(MarketingDataCleanGeneralConfigExample example);
 
     int deleteByExample(MarketingDataCleanGeneralConfigExample example);
